pyproject.toml
[project]
name = "project-rh"
version = "0.1.0"
description = ""
authors = [
    {name = "Ailer20",email = "aronvieira820@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (==5.2.7)",
    "djangorestframework (==3.16.1)",
    "python-decouple (==3.8)",
    "pillow (==12.0.0)",
    "asgiref (==3.10.0)",
    "sqlparse (==0.5.3)",
    "django-import-export (>=4.3.12,<5.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
